Refined add_item & is_stackable
In add item the first check to see if the inventory is full also checks to see if the item can be stacked. I do this because when the inventory reaches the last slot and there's 1 item in the stack it thinks the inventory is full.

is_stackable() now has something in it to complete the first check with is_full()